pyproject.toml
[project]
name = "lohg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "keyboard>=0.13.5",
    "mouse>=0.7.1",
    "mss>=10.0.0",
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "pylance>=0.23.2",
    "pynput>=1.7.7",
    "pytesseract>=0.3.13",
    "rich>=13.9.4",
    "ruff>=0.9.7",
]

[dependency-groups]
dev = [
    "ruff>=0.9.7",
]
